Move Exam and StudentExam to domain package
