+++ src/main/java/de/tum/in/www1/artemis/domain/ExerciseGroup.java
@@ -15,8 +15,6 @@ import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
-import de.tum.in.www1.artemis.domain.exam.Exam;
-
 @Entity
 @Table(name = "exercise_group")
 public class ExerciseGroup {
